Avancement de run.py et création d'install.py

+++ BLOCKCHAINEXPLORER/run.py
@@ -1,25 +1,32 @@
 import os, shutil, glob
 import subprocess, time, webbrowser
+import platform
 
-pathGlob = glob.glob('/home/alexandre/Projet-L3AB1/BLOCKCHAINEXPLORER/datas/*')
-if(os.path.exists('/home/alexandre/Projet-L3AB1/BLOCKCHAINEXPLORER/datas')):
-   for f in pathGlob:
-      if(os.path.isfile(f)):
-         os.remove(f)
-      if(os.path.isdir(f)):
-         shutil.rmtree(f)
-processMongoDB = subprocess.Popen(["mongod", "--dbpath", "datas/"])
-processDjango = subprocess.Popen(["python3", "manage.py", "runserver"])
-time.sleep(8)
-webbrowser.open('http://127.0.0.1:8000/')
-time.sleep(5)
-close = int(input("Si vous voulez fermer le site, tapez 1"))
-if(close == 1):
-   processMongoDB.kill()
-   processDjango.kill()
-   if(os.path.exists('/home/alexandre/Projet-L3AB1/BLOCKCHAINEXPLORER/datas')):
+monOS = platform.system()
+if(monOS == 'Linux' or monOS == 'Darwin'):
+   pathGlob = glob.glob('datas/*')
+   if(os.path.exists('datas')):
       for f in pathGlob:
          if(os.path.isfile(f)):
             os.remove(f)
          if(os.path.isdir(f)):
             shutil.rmtree(f)
+   processMongoDB = subprocess.Popen(["mongod", "--dbpath", "datas/"])
+   processDjango = subprocess.Popen(["python3", "manage.py", "runserver"])
+   time.sleep(8)
+   webbrowser.open('http://127.0.0.1:8000/')
+   close = 0
+   while(close != 1):
+      time.sleep(5)
+      close = int(input("Si vous voulez fermer le site, tapez 1\n"))
+   if(close == 1):
+      processMongoDB.kill()
+      processDjango.kill()
+      if(os.path.exists('datas')):
+         for f in pathGlob:
+            if(os.path.isfile(f)):
+               os.remove(f)
+            if(os.path.isdir(f)):
+               shutil.rmtree(f)
+else:
+   print("je suis sur windows")
